App jsx refactor
